Syri batch now can split fasta and align each chr, too

+++ wrapper/iga/apps/base.py
@@ -86,6 +86,8 @@ def split_fasta(fasta, workdir, chunk=100, bypart='F'):
             for i, j in enumerate(data[1:], start=1):
                 (header, content) = j.split('\n', 1)
                 new_file_name = f'{header}.fasta'
+                if os.path.exists(new_file_name):
+                    continue
                 file_list.append(new_file_name)
                 with open(new_file_name, 'w') as f:
                     f.write('>' + j)
